+++ Modules/Microsoft365DSC/DscResources/MSFT_AADAccessReviewDefinition/MSFT_AADAccessReviewDefinition.psm1
@@ -1178,6 +1178,7 @@ function Export-TargetResource
         [array]$getValue = Get-MgBetaIdentityGovernanceAccessReviewDefinition `
             -Filter $Filter `
             -All `
+            -Top 100 `
             -ErrorAction Stop
         #endregion