+++ Modules/Microsoft365DSC/DscResources/MSFT_AADAuthenticationContextClassReference/MSFT_AADAuthenticationContextClassReference.psm1
@@ -415,6 +415,7 @@ function Export-TargetResource
         [array]$getValue = Get-MgBetaIdentityConditionalAccessAuthenticationContextClassReference `
             -All `
             -Filter $Filter `
+            -Top 0 `
             -ErrorAction Stop
         #endregion