+++ Utilities/New-M365DSCGraphShimModule.ps1
@@ -336,6 +336,14 @@ function Invoke-M365DSCGraphShimRequestV75
         $OutputType,
 
         [Parameter()]
+        [System.Int32]
+        $Skip = 0,
+
+        [Parameter()]
+        [System.Int32]
+        $Top = 0,
+
+        [Parameter()]
         [switch]
         $All,
 
@@ -355,6 +363,20 @@ function Invoke-M365DSCGraphShimRequestV75
         Method     = $Method
         Uri        = $Uri.TrimStart('/beta').TrimStart('/v1.0')
         ErrorAction = 'Stop'
+    }
+
+    if ($PSBoundParameters.ContainsKey('Skip') -and $Skip -gt 0)
+    {
+        $invokeParams['Skip'] = $Skip
+    }
+
+    if ($PSBoundParameters.ContainsKey('Top') -and $Top -gt 0)
+    {
+        $invokeParams['PageSize'] = $Top
+    }
+    elseif ($PSBoundParameters.ContainsKey('Top') -and $Top -eq 0)
+    {
+        $invokeParams['NoPageSize'] = $true
     }
 
     if ($PSBoundParameters.ContainsKey('Body') -and $null -ne $Body)
@@ -421,11 +443,21 @@ function Get-M365DSCGraphShimAllPages
 
         [Parameter()]
         [System.Int32]
+        $Skip = 0,
+
+        [Parameter()]
+        [System.Int32]
         $Top = 0
     )
 
     $allResults = [System.Collections.Generic.List[System.Object]]::new()
     $currentUri = $Uri
+
+    if ($Skip -gt 0 -and $currentUri -notmatch '[\?&]\$skip=')
+    {
+        $separator = if ($currentUri.Contains('?')) { '&' } else { '?' }
+        $currentUri = "$currentUri$separator`$skip=$Skip"
+    }
 
     if ($Top -gt 0 -and $currentUri -notmatch '[\?&]\$top=')
     {
@@ -488,26 +520,35 @@ function Get-M365DSCGraphShimAllPagesV75
 
         [Parameter()]
         [System.Int32]
+        $Skip = 0,
+
+        [Parameter()]
+        [System.Int32]
         $Top = 0
     )
 
     $allResults = [System.Collections.Generic.List[System.Object]]::new()
     $currentUri = $Uri
 
-    if ($Top -gt 0 -and $currentUri -notmatch '[\?&]\$top=')
-    {
-        $separator = if ($currentUri.Contains('?')) { '&' } else { '?' }
-        $currentUri = "$currentUri$separator`$top=$Top"
-    }
-
     $requestParams = @{
         All    = $true
         Method = 'GET'
         Uri    = $currentUri
     }
+
     if ($PSBoundParameters.ContainsKey('Headers') -and $Headers.Keys.Count -gt 0)
     {
         $requestParams['Headers'] = $Headers
+    }
+
+    if ($PSBoundParameters.ContainsKey('Skip'))
+    {
+        $requestParams['Skip'] = $Skip
+    }
+
+    if ($PSBoundParameters.ContainsKey('Top'))
+    {
+        $requestParams['Top'] = $Top
     }
 
     $response = Invoke-M365DSCGraphShimRequestV75 @requestParams -PassThru
@@ -742,11 +783,13 @@ function Invoke-M365DSCGraphShimGetResource
     if ($BoundParameters['Filter'])         { $uriParams['Filter'] = $BoundParameters['Filter'] }
     if ($BoundParameters['Property'])       { $uriParams['Property'] = $BoundParameters['Property'] }
     if ($BoundParameters['ExpandProperty']) { $uriParams['ExpandProperty'] = $BoundParameters['ExpandProperty'] }
-    if ($BoundParameters.ContainsKey('Top') -and $BoundParameters['Top'] -gt 0)   { $uriParams['Top'] = $BoundParameters['Top'] }
-    if ($BoundParameters.ContainsKey('Skip') -and $BoundParameters['Skip'] -gt 0) { $uriParams['Skip'] = $BoundParameters['Skip'] }
     if ($BoundParameters['Search'])         { $uriParams['Search'] = $BoundParameters['Search'] }
     if ($BoundParameters['Sort'])           { $uriParams['Sort'] = $BoundParameters['Sort'] }
     if ($BoundParameters['CountVariable'])  { $uriParams['CountVariable'] = $BoundParameters['CountVariable'] }
+
+    $paramSplat = @{}
+    if ($BoundParameters.ContainsKey('Top'))   { $paramSplat['Top'] = $BoundParameters['Top'] }
+    if ($BoundParameters.ContainsKey('Skip')) { $paramSplat['Skip'] = $BoundParameters['Skip'] }
 
     $uri = ConvertTo-M365DSCGraphShimUri @uriParams
 
@@ -754,9 +797,9 @@ function Invoke-M365DSCGraphShimGetResource
     {
         if ($Script:IsPowerShell75OrGreater)
         {
-            return Get-M365DSCGraphShimAllPagesV75 -Uri $uri -Headers $requestHeaders -ErrorAction $ErrorActionPreference
+            return Get-M365DSCGraphShimAllPagesV75 -Uri $uri -Headers $requestHeaders @paramSplat -ErrorAction $ErrorActionPreference
         }
-        return Get-M365DSCGraphShimAllPages -Uri $uri -Headers $requestHeaders -ErrorAction $ErrorActionPreference
+        return Get-M365DSCGraphShimAllPages -Uri $uri -Headers $requestHeaders @paramSplat -ErrorAction $ErrorActionPreference
     }
 
     if ($Script:IsPowerShell75OrGreater)